A more compatible project / workspace with CocoaPods. Usint pod lib lint SweetAlert.podspec to check the pod Creating the Example folder.
