+++ Example/SweetAlert/AppDelegate.swift
