+++ Example/SweetAlert/Base.lproj/LaunchScreen.xib
