+++ Example/SweetAlert/Info.plist
