+++ Example/SweetAlert/ViewController.swift
@@ -7,6 +7,7 @@
 //
 
 import UIKit
+import SweetAlert
 
 class ViewController: UIViewController {
